chore: refactor lab store
